Multi-Dimensional Signed Singular Value Polyconvexification

API

NumericalRelaxation.PolyConvexification — Type

PolyConvexification{} <: AbstractConvexification

Datastructure which holds basic parameters and grid for the Polyconvexification

dimp::Int dimension of the physical problem

dimc::Int lifted dimension, 2D -> 3, 3D -> 7, in this dimension the convexificaiton problem for the polyconvexification is stated

r::Float64 discretization radius

nref::Int number of uniform grid refinements

grid::Vector{T1} grid of the signed singula values

liftedGrid::Vector{T1} lifted grid of signed singular values through application of the minors function

NumericalRelaxation.PolyConvexificationBuffer — Type

Φν_δ::Vector{T1} holds the values of Φ evaluated at the grid ν_δ

Φactive::Vector{Bool} marks the grid points involved in the minimization problem

NumericalRelaxation.convexify — Method

convexify(poly_convexification::PolyConvexification, poly_buffer::PolyConvexificationBuffer, Φ::FUN, ν::Union{Vec{d},Vector{Float64}}, xargs::Vararg{Any,XN}; returnDerivs::Bool=true) where {FUN,XN,d}

Signed singular value polyconvexification using the linear programming approach. Compute approximation to the singular value polycovex envelope of the function Φ which is the reformulation of the isotropic function W in terms of signed singular values $Φ(ν) = W(diagm(ν))$, at the point ν via the linear programming approach as discussed in [1] Timo Neumeier, Malte A. Peter, Daniel Peterseim, David Wiedemann. Computational polyconvexification of isotropic functions, arXiv 2307.15676, 2023. The parameters nref and r (stored in poly_convexification struct) discribe the grid by radius r (in the ∞ norm) and nref uniform mesh refinements. The points of the lifted grid which are involved in the minimization are marked by the Φactive buffer, and deliver Φ values smaller than infinity.

Φ::FUN function in terms of signed singular values Φ(ν) = W(diagm(ν)) ν::Vector{Float64} point of evaluation for the polyconvex hull returnDerivs::Bool return first order derivative information

NumericalRelaxation.ssv — Function

signed singular values

NumericalRelaxation.Dssv — Function

Derivative of signed singular values mapping, calculated by forward difference quotients. TODO: might need improvement

NumericalRelaxation.minors — Function

minors for matrices and vectors

NumericalRelaxation.Dminors — Function

Derivative of the minors function Dminors$:\mathbb{R}^{d} \to \mathbb{R}^{d \times k_d}$ with $k_d$ denoting the lifted dimension