Multi-Dimensional Signed Singular Value Polyconvexification

API

NumericalRelaxation.PolyConvexification — Type

PolyConvexification{} <: AbstractConvexification

Datastructure which holds basic parameters and grid for the Polyconvexification

dimp::Int dimension of the physical problem

dimc::Int lifted dimension, 2D -> 3, 3D -> 7, in this dimension the convexificaiton problem for the polyconvexification is stated

r::Float64 discretization radius

nref::Int number of uniform grid refinements

grid::Vector{T1} grid of the signed singula values

liftedGrid::Vector{T1} lifted grid of signed singular values through application of the minors function

NumericalRelaxation.PolyConvexificationBuffer — Type

Φν_δ::Vector{T1} holds the values of Φ evaluated at the grid ν_δ

Φactive::Vector{Bool} marks the grid points involved in the minimization problem

Missing docstring.

Missing docstring for NumericalRelaxation.convexify. Check Documenter's build log for details.

NumericalRelaxation.ssv — Function

signed singular values

NumericalRelaxation.Dssv — Function

Derivative of signed singular values mapping, calculated by forward difference quotients. TODO: might need improvement

NumericalRelaxation.minors — Function

minors for matrices and vectors

NumericalRelaxation.Dminors — Function

Derivative of the minors function Dminors$:\mathbb{R}^{d} \to \mathbb{R}^{d \times k_d}$ with $k_d$ denoting the lifted dimension